% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{createPlot}
\alias{createPlot}
\title{Create ROI posterior distribution plot}
\usage{
createPlot(a, b, c, d)
}
\arguments{
\item{a}{wide dataframe of posteriors (character)}

\item{b}{title of plot (character)}

\item{c}{right-most (negative) x-axis label}

\item{d}{left-most (positive) x-axis label}
}
\value{
\strong{df_long.csv} exports long dataframe with P+ values alongside posteriors and means

\strong{P.csv} exports P+ summary table

\strong{TITLE.png} high resolution posterior plot (dpi = 800, height = 7, width = 9)
}
\description{
Summarize region of interest (ROI) outputs of a Bayesian model using a ridgeline posterior plot. Each ROI's posterior distribution is displayed along with their P+ values (see Limbachia et al, 2020 for details) \cr
This summary & plot are generated from wideform .txt or .csv data. \cr
For now this package only works with files that has no separator/delimiter. It is best if you have pretty ROI names ahead of time, this package will print whatever the column values are from the input dataset.
}
\details{
\strong{This package will:} \cr
\enumerate{
\item Transpose wideform dataset \cr
a. columns are ROIs, rows are posterior values \cr
\item Calculate P+ values \cr
\item Output table of P+ values per ROI \cr
\item Generate standardized ridgeline posterior plot \cr
}

\strong{ALL PARAMS SHOULD ENTERED IN QUOTES BECAUSE THEY ARE CHARACTERS!} \cr
Example: \emph{createPlot("path/to/file.txt", "Plot title", "X > Y", "Y > X")} \cr
If you do not want x-axis labels, enter NA for both c and d values. \cr
}
