% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlot.R
\name{createPlot}
\alias{createPlot}
\title{Create posterior plot}
\usage{
createPlot(a, b, c, d)
}
\arguments{
\item{a}{wide dataframe of posteriors}

\item{b}{title of plot}

\item{c}{right-most (negative) x-axis label}

\item{d}{left-most (positive) x-axis label}
}
\value{
df exports long dataframe with P+ values

df exports P summary table

high resolution posterior plot (dpi = 800, height = 7, width = 9)
}
\description{
Generate a posterior ridgeline plot from wideframe text .txt or .csv
(cols = ROIs, rows = posteriors).
}
\details{
ALL PARAMS SHOULD ENTERED IN QUOTES BECAUSE THEY ARE CHARACTERS!
Example: createPlot("path/to/file.txt", "Plot title", "X > Y", "Y > X")
If you do not want X-axis labels, enter NULL for both c and d values.

For now this package only works with files that has no separator ""
It is best if you have pretty ROI names ahead of time, this package will print whatever the column values are from the input dataset.
}
